Refactor MACE wrapper helpers
